requirements.txt
streamlit
openai
python-dotenv
diffusers
transformers
scipy
ftfy
accelerate
 midjourney-api
